refactor: split _supabase.py god module into focused helpers (#402)
Moves notification and sync-log concerns into dedicated modules to
reduce blast radius and enable isolated testing of each concern.
