+++ scripts/check_daily_alerts.py
@@ -18,7 +18,8 @@ from pathlib import Path
 
 ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(ROOT / "scripts"))
-from _supabase import get_client, get_owner_user_id, log_notification
+from _supabase import get_client, get_owner_user_id
+from _notifications import log_notification
 
 NOTIFY_SCRIPT = ROOT / "scripts" / "notify.sh"
 CLICK_PATH = "/tasks"
